+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix `install-apex-executables` script hanging unnecessarily ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/1368)).
 - Prettier Apex should now throw an error with details when it fails to call Apex AST Serializer executable.
 - NPM package tests are now run on multiple platforms to catch any platform-dependent regressions.
+- Use relative path to binary files to avoid issues with special characters in Windows path.
 
 # 2.1.3